+++ predicators/ground_truth_models/__init__.py
@@ -129,9 +129,7 @@ utils.import_submodules(__path__, __name__)
 
 def deprecated_get_gt_nsrts(env_name: str) -> Set[NSRT]:
     """Create ground truth NSRTs for an env."""
-    if env_name == "coffee":
-        nsrts = _get_coffee_gt_nsrts(env_name)
-    elif env_name in ("satellites", "satellites_simple"):
+    if env_name in ("satellites", "satellites_simple"):
         nsrts = _get_satellites_gt_nsrts(env_name)
     else:
         assert env_name in ("sandwich", "sandwich_clear")
@@ -167,292 +165,6 @@ def _get_options_by_names(env_name: str,
     options = get_gt_options(env_name)
     name_to_option = {o.name: o for o in options}
     return [name_to_option[name] for name in names]
-
-
-def _get_narrow_passage_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for NarrowPassageEnv."""
-    robot_type, door_type, target_type = _get_types_by_names(
-        env_name, ["robot", "door", "target"])
-    DoorIsClosed, DoorIsOpen, TouchedGoal = _get_predicates_by_names(
-        env_name, ["DoorIsClosed", "DoorIsOpen", "TouchedGoal"])
-    MoveToTarget, MoveAndOpenDoor = _get_options_by_names(
-        env_name, ["MoveToTarget", "MoveAndOpenDoor"])
-
-    nsrts = set()
-
-    def random_sampler(state: State, goal: Set[GroundAtom],
-                       rng: np.random.Generator,
-                       objs: Sequence[Object]) -> Array:
-        del state, goal, objs  # unused
-        # Note: just return a random value from 0 to 1
-        return np.array([rng.uniform()], dtype=np.float32)
-
-    # MoveToTarget
-    robot = Variable("?robot", robot_type)
-    target = Variable("?target", target_type)
-    parameters = [robot, target]
-    option_vars = [robot, target]
-    option = MoveToTarget
-    preconditions: Set[LiftedAtom] = set()
-    add_effects: Set[LiftedAtom] = {
-        LiftedAtom(TouchedGoal, [robot, target]),
-    }
-    delete_effects: Set[LiftedAtom] = set()
-    ignore_effects: Set[Predicate] = set()
-    move_to_target_nsrt = NSRT("MoveToTarget", parameters, preconditions,
-                               add_effects, delete_effects, ignore_effects,
-                               option, option_vars, random_sampler)
-    nsrts.add(move_to_target_nsrt)
-
-    # MoveAndOpenDoor
-    robot = Variable("?robot", robot_type)
-    door = Variable("?door", door_type)
-    parameters = [robot, door]
-    option_vars = [robot, door]
-    option = MoveAndOpenDoor
-    preconditions = {
-        LiftedAtom(DoorIsClosed, [door]),
-    }
-    add_effects = {
-        LiftedAtom(DoorIsOpen, [door]),
-    }
-    delete_effects = {
-        LiftedAtom(DoorIsClosed, [door]),
-    }
-    ignore_effects = set()
-    move_and_open_door_nsrt = NSRT("MoveAndOpenDoor", parameters,
-                                   preconditions, add_effects, delete_effects,
-                                   ignore_effects, option, option_vars,
-                                   random_sampler)
-    nsrts.add(move_and_open_door_nsrt)
-
-    return nsrts
-
-
-def _get_coffee_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for CoffeeEnv."""
-    robot_type, jug_type, cup_type, machine_type = _get_types_by_names(
-        env_name, ["robot", "jug", "cup", "machine"])
-    CupFilled, Holding, JugInMachine, MachineOn, OnTable, HandEmpty, \
-        JugFilled, RobotAboveCup, JugAboveCup, NotAboveCup, PressingButton, \
-        Twisting, NotSameCup = \
-        _get_predicates_by_names(env_name, ["CupFilled",
-            "Holding", "JugInMachine", "MachineOn", "OnTable", "HandEmpty",
-            "JugFilled", "RobotAboveCup", "JugAboveCup", "NotAboveCup",
-            "PressingButton", "Twisting", "NotSameCup"])
-    MoveToTwistJug, TwistJug, PickJug, PlaceJugInMachine, TurnMachineOn, \
-        Pour = _get_options_by_names(env_name, ["MoveToTwistJug", "TwistJug",
-            "PickJug", "PlaceJugInMachine", "TurnMachineOn", "Pour"])
-
-    nsrts = set()
-
-    # MoveToTwistJug
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    parameters = [robot, jug]
-    option_vars = [robot, jug]
-    option = MoveToTwistJug
-    preconditions = {
-        LiftedAtom(OnTable, [jug]),
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    add_effects = {
-        LiftedAtom(Twisting, [robot, jug]),
-    }
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    ignore_effects: Set[Predicate] = set()
-    move_to_twist_jug_nsrt = NSRT("MoveToTwistJug", parameters, preconditions,
-                                  add_effects, delete_effects, ignore_effects,
-                                  option, option_vars, null_sampler)
-    nsrts.add(move_to_twist_jug_nsrt)
-
-    # TwistJug
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    parameters = [robot, jug]
-    option_vars = [robot, jug]
-    option = TwistJug
-    preconditions = {
-        LiftedAtom(OnTable, [jug]),
-        LiftedAtom(Twisting, [robot, jug]),
-    }
-    add_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    delete_effects = {
-        LiftedAtom(Twisting, [robot, jug]),
-    }
-    ignore_effects = set()
-
-    def twist_jug_sampler(state: State, goal: Set[GroundAtom],
-                          rng: np.random.Generator,
-                          objs: Sequence[Object]) -> Array:
-        del state, goal, objs  # unused
-        return np.array(rng.uniform(-1, 1, size=(1, )), dtype=np.float32)
-
-    twist_jug_nsrt = NSRT("TwistJug", parameters, preconditions, add_effects,
-                          delete_effects, ignore_effects, option, option_vars,
-                          twist_jug_sampler)
-    nsrts.add(twist_jug_nsrt)
-
-    # PickJugFromTable
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    parameters = [robot, jug]
-    option_vars = [robot, jug]
-    option = PickJug
-    preconditions = {
-        LiftedAtom(OnTable, [jug]),
-        LiftedAtom(HandEmpty, [robot])
-    }
-    add_effects = {
-        LiftedAtom(Holding, [robot, jug]),
-    }
-    delete_effects = {
-        LiftedAtom(OnTable, [jug]),
-        LiftedAtom(HandEmpty, [robot])
-    }
-    ignore_effects = set()
-    pick_jug_from_table_nsrt = NSRT("PickJugFromTable", parameters,
-                                    preconditions, add_effects, delete_effects,
-                                    ignore_effects, option, option_vars,
-                                    null_sampler)
-    nsrts.add(pick_jug_from_table_nsrt)
-
-    # PlaceJugInMachine
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    machine = Variable("?machine", machine_type)
-    parameters = [robot, jug, machine]
-    option_vars = [robot, jug, machine]
-    option = PlaceJugInMachine
-    preconditions = {
-        LiftedAtom(Holding, [robot, jug]),
-    }
-    add_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(JugInMachine, [jug, machine]),
-    }
-    delete_effects = {
-        LiftedAtom(Holding, [robot, jug]),
-    }
-    ignore_effects = set()
-    place_jug_in_machine_nsrt = NSRT("PlaceJugInMachine", parameters,
-                                     preconditions, add_effects,
-                                     delete_effects, ignore_effects, option,
-                                     option_vars, null_sampler)
-    nsrts.add(place_jug_in_machine_nsrt)
-
-    # TurnMachineOn
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    machine = Variable("?machine", machine_type)
-    parameters = [robot, jug, machine]
-    option_vars = [robot, machine]
-    option = TurnMachineOn
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(JugInMachine, [jug, machine]),
-    }
-    add_effects = {
-        LiftedAtom(JugFilled, [jug]),
-        LiftedAtom(MachineOn, [machine]),
-        LiftedAtom(PressingButton, [robot, machine]),
-    }
-    delete_effects = set()
-    ignore_effects = set()
-    turn_machine_on_nsrt = NSRT("TurnMachineOn", parameters, preconditions,
-                                add_effects, delete_effects, ignore_effects,
-                                option, option_vars, null_sampler)
-    nsrts.add(turn_machine_on_nsrt)
-
-    # PickJugFromMachine
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    machine = Variable("?machine", machine_type)
-    parameters = [robot, jug, machine]
-    option_vars = [robot, jug]
-    option = PickJug
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(JugInMachine, [jug, machine]),
-        LiftedAtom(PressingButton, [robot, machine]),
-    }
-    add_effects = {
-        LiftedAtom(Holding, [robot, jug]),
-    }
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(JugInMachine, [jug, machine]),
-        LiftedAtom(PressingButton, [robot, machine]),
-    }
-    ignore_effects = set()
-    pick_jug_from_machine_nsrt = NSRT("PickJugFromMachine", parameters,
-                                      preconditions, add_effects,
-                                      delete_effects, ignore_effects, option,
-                                      option_vars, null_sampler)
-    nsrts.add(pick_jug_from_machine_nsrt)
-
-    # PourFromNowhere
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    cup = Variable("?cup", cup_type)
-    parameters = [robot, jug, cup]
-    option_vars = [robot, jug, cup]
-    option = Pour
-    preconditions = {
-        LiftedAtom(Holding, [robot, jug]),
-        LiftedAtom(JugFilled, [jug]),
-        LiftedAtom(NotAboveCup, [robot, jug]),
-    }
-    add_effects = {
-        LiftedAtom(JugAboveCup, [jug, cup]),
-        LiftedAtom(RobotAboveCup, [robot, cup]),
-        LiftedAtom(CupFilled, [cup]),
-    }
-    delete_effects = {
-        LiftedAtom(NotAboveCup, [robot, jug]),
-    }
-    ignore_effects = set()
-    pour_from_nowhere_nsrt = NSRT("PourFromNowhere", parameters, preconditions,
-                                  add_effects, delete_effects, ignore_effects,
-                                  option, option_vars, null_sampler)
-    nsrts.add(pour_from_nowhere_nsrt)
-
-    # PourFromOtherCup
-    robot = Variable("?robot", robot_type)
-    jug = Variable("?jug", jug_type)
-    cup = Variable("?cup", cup_type)
-    other_cup = Variable("?other_cup", cup_type)
-    parameters = [robot, jug, cup, other_cup]
-    option_vars = [robot, jug, cup]
-    option = Pour
-    preconditions = {
-        LiftedAtom(Holding, [robot, jug]),
-        LiftedAtom(JugFilled, [jug]),
-        LiftedAtom(JugAboveCup, [jug, other_cup]),
-        LiftedAtom(RobotAboveCup, [robot, other_cup]),
-        LiftedAtom(NotSameCup, [cup, other_cup]),
-    }
-    add_effects = {
-        LiftedAtom(JugAboveCup, [jug, cup]),
-        LiftedAtom(RobotAboveCup, [robot, cup]),
-        LiftedAtom(CupFilled, [cup]),
-    }
-    delete_effects = {
-        LiftedAtom(JugAboveCup, [jug, other_cup]),
-        LiftedAtom(RobotAboveCup, [robot, other_cup]),
-    }
-    ignore_effects = set()
-    pour_from_other_cup_nsrt = NSRT("PourFromOtherCup", parameters,
-                                    preconditions, add_effects, delete_effects,
-                                    ignore_effects, option, option_vars,
-                                    null_sampler)
-    nsrts.add(pour_from_other_cup_nsrt)
-
-    return nsrts
 
 
 def _get_satellites_gt_nsrts(env_name: str) -> Set[NSRT]:
