+++ predicators/ground_truth_models/coffee/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for coffee environment and variants."""
+
+from .nsrts import CoffeeGroundTruthNSRTFactory
+
+__all__ = ["CoffeeGroundTruthNSRTFactory"]
